# php-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice Management API project",
        "description": "Invoice Management API project. It provides endpoints to manage users, tax profiles, and invoices. The project is structured using the MVC architecture and includes automated tests, Docker support, and API documentation via Postman.",
        "contact": {
            "email": "caiazza.domenico@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/users/{userId}/tax-profiles/{taxProfileId}/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Retrieve a list of invoices",
                "description": "Returns a paginated list of invoices, with the ability to filter by invoice number, date, currency, status, and notes.",
                "operationId": "a3982b788f865f5eb109853f9431c83c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "taxProfileId",
                        "in": "path",
                        "description": "ID of the tax profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[invoice_number]",
                        "in": "query",
                        "description": "Filter by invoice number (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[invoice_date]",
                        "in": "query",
                        "description": "Filter by invoice date (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "filter[invoice_date_before]",
                        "in": "query",
                        "description": "Filter invoices with a date less than or equal to the specified date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "filter[invoice_date_after]",
                        "in": "query",
                        "description": "Filter invoices with a date greater than or equal to the specified date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "filter[currency]",
                        "in": "query",
                        "description": "Filter by currency",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[notes]",
                        "in": "query",
                        "description": "Filter by notes (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Invoice"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Links"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Meta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "description": "Creates a new invoice for the specified tax profile.",
                "operationId": "createInvoice",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "taxProfileId",
                        "in": "path",
                        "description": "ID of the tax profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_date",
                                    "subtotal",
                                    "tax_amount",
                                    "discount"
                                ],
                                "properties": {
                                    "invoice_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-01"
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "tax_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.07
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "EUR"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Invoice for services rendered"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Invoice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Invoice": {
                "required": [
                    "id",
                    "user_id",
                    "tax_profile_id",
                    "subtotal",
                    "tax_amount",
                    "invoice_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tax_profile_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_number": {
                        "type": "string",
                        "example": "1740839386_NMEPW"
                    },
                    "invoice_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2021-01-01"
                    },
                    "subtotal": {
                        "type": "string",
                        "example": "100.00"
                    },
                    "tax_amount": {
                        "type": "string",
                        "example": "22.00"
                    },
                    "discount": {
                        "type": "string",
                        "example": "00.00"
                    },
                    "currency": {
                        "type": "string",
                        "example": "EUR"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null"
                    },
                    "canceled_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Comment"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Record not found"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Results per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Results page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "xApiKeyHeader": {
                "name": "x-api-key",
                "in": "header",
                "description": "API key authentication",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}